Introduce a small abstraction to verify a session
